+++ app/core/Engine.ts
@@ -21,7 +21,11 @@ import {
   TokenRatesState,
   TokensController,
   TokensState,
+  CodefiTokenPricesServiceV2,
 } from '@metamask/assets-controllers';
+///: BEGIN:ONLY_INCLUDE_IF(snaps)
+import { AppState } from 'react-native';
+///: END:ONLY_INCLUDE_IF
 import {
   AddressBookController,
   AddressBookState,
@@ -34,6 +38,9 @@ import {
   KeyringControllerState,
   KeyringControllerActions,
   KeyringControllerEvents,
+  ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+  KeyringTypes,
+  ///: END:ONLY_INCLUDE_IF
 } from '@metamask/keyring-controller';
 import {
   NetworkController,
@@ -72,16 +79,45 @@ import {
   PermissionControllerActions,
   PermissionControllerEvents,
   PermissionControllerState,
+  ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+  SubjectMetadataController,
+  SubjectMetadataControllerActions,
+  SubjectMetadataControllerEvents,
+  SubjectMetadataControllerState,
+  SubjectType,
+  ///: END:ONLY_INCLUDE_IF
 } from '@metamask/permission-controller';
 import SwapsController, { swapsUtils } from '@metamask/swaps-controller';
-import { PPOMController, PPOMState } from '@metamask/ppom-validator';
+import {
+  PPOMController,
+  PPOMControllerEvents,
+  PPOMInitialisationStatusType,
+  PPOMState,
+} from '@metamask/ppom-validator';
+///: BEGIN:ONLY_INCLUDE_IF(snaps)
+import {
+  JsonSnapsRegistry,
+  AllowedActions as SnapsAllowedActions,
+  AllowedEvents as SnapsAllowedEvents,
+  SnapController,
+  SnapsRegistryState,
+  SnapControllerEvents,
+  SnapControllerActions,
+  buildSnapEndowmentSpecifications,
+  buildSnapRestrictedMethodSpecifications,
+  PersistedSnapControllerState,
+} from '@metamask/snaps-controllers';
+import { EnumToUnion, Snap } from '@metamask/snaps-utils';
+import { DialogType, NotificationArgs } from '@metamask/snaps-rpc-methods';
+// eslint-disable-next-line import/no-nodejs-modules
+import { Duplex } from 'stream';
+///: END:ONLY_INCLUDE_IF
 import { MetaMaskKeyring as QRHardwareKeyring } from '@keystonehq/metamask-airgapped-keyring';
 import {
   LoggingController,
   LoggingControllerState,
   LoggingControllerActions,
 } from '@metamask/logging-controller';
-import LedgerKeyring from '@consensys/ledgerhq-metamask-keyring';
 import Encryptor from './Encryptor';
 import {
   isMainnetByChainId,
@@ -99,9 +135,24 @@ import {
 } from '../util/number';
 import NotificationManager from './NotificationManager';
 import Logger from '../util/Logger';
+///: BEGIN:ONLY_INCLUDE_IF(snaps)
+import { EndowmentPermissions } from '../constants/permissions';
+///: END:ONLY_INCLUDE_IF
 import { isZero } from '../util/lodash';
 import { MetaMetricsEvents } from './Analytics';
 import AnalyticsV2 from '../util/analyticsV2';
+///: BEGIN:ONLY_INCLUDE_IF(snaps)
+import {
+  SnapBridge,
+  WebviewExecutionService,
+  ExcludedSnapEndowments,
+  ExcludedSnapPermissions,
+  detectSnapLocation,
+  fetchFunction,
+  DetectSnapLocationOptions,
+} from './Snaps';
+import { getRpcMethodMiddleware } from './RPCMethods/RPCMethodMiddleware';
+///: END:ONLY_INCLUDE_IF
 import { isBlockaidFeatureEnabled } from '../util/blockaid';
 import {
   getCaveatSpecifications,
@@ -124,11 +175,37 @@ import RNFSStorageBackend from '../lib/ppom/rnfs-storage-backend';
 import { isHardwareAccount } from '../util/address';
 import { ledgerSignTypedMessage } from './Ledger/Ledger';
 import ExtendedKeyringTypes from '../constants/keyringTypes';
+import { UpdatePPOMInitializationStatus } from '../actions/experimental';
 
 const NON_EMPTY = 'NON_EMPTY';
 
 const encryptor = new Encryptor();
 let currentChainId: any;
+
+///: BEGIN:ONLY_INCLUDE_IF(snaps)
+// TODO remove these custom types when the PhishingController is to version >= 7.0.0
+interface MaybeUpdateState {
+  type: `${PhishingController['name']}:maybeUpdateState`;
+  handler: PhishingController['maybeUpdateState'];
+}
+
+interface TestOrigin {
+  type: `${PhishingController['name']}:testOrigin`;
+  handler: PhishingController['test'];
+}
+type PhishingControllerActions = MaybeUpdateState | TestOrigin;
+
+type SnapsGlobalActions =
+  | SnapControllerActions
+  | SubjectMetadataControllerActions
+  | PhishingControllerActions
+  | SnapsAllowedActions;
+
+type SnapsGlobalEvents =
+  | SnapControllerEvents
+  | SubjectMetadataControllerEvents
+  | SnapsAllowedEvents;
+///: END:ONLY_INCLUDE_IF
 
 type GlobalActions =
   | ApprovalControllerActions
@@ -138,8 +215,11 @@ type GlobalActions =
   | NetworkControllerActions
   | PermissionControllerActions
   | SignatureControllerActions
-  | KeyringControllerActions
-  | LoggingControllerActions;
+  | LoggingControllerActions
+  ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+  | SnapsGlobalActions
+  ///: END:ONLY_INCLUDE_IF
+  | KeyringControllerActions;
 type GlobalEvents =
   | ApprovalControllerEvents
   | CurrencyRateStateChange
@@ -147,8 +227,12 @@ type GlobalEvents =
   | TokenListStateChange
   | NetworkControllerEvents
   | PermissionControllerEvents
+  | KeyringControllerEvents
+  ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+  | SnapsGlobalEvents
+  ///: END:ONLY_INCLUDE_IF
   | SignatureControllerEvents
-  | KeyringControllerEvents;
+  | PPOMControllerEvents;
 
 type PermissionsByRpcMethod = ReturnType<typeof getPermissionSpecifications>;
 type Permissions = PermissionsByRpcMethod[keyof PermissionsByRpcMethod];
@@ -172,6 +256,11 @@ export interface EngineState {
   TokensController: TokensState;
   TokenDetectionController: BaseState;
   NftDetectionController: BaseState;
+  ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+  SnapController: PersistedSnapControllerState;
+  SnapsRegistry: SnapsRegistryState;
+  SubjectMetadataController: SubjectMetadataControllerState;
+  ///: END:ONLY_INCLUDE_IF
   PermissionController: PermissionControllerState<Permissions>;
   ApprovalController: ApprovalControllerState;
   LoggingController: LoggingControllerState;
@@ -232,6 +321,13 @@ class Engine {
    * for each address and network
    */
   lastIncomingTxBlockInfo: any;
+
+  ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+  /**
+   * Object that runs and manages the execution of Snaps
+   */
+  snapExecutionService: WebviewExecutionService;
+  ///: END:ONLY_INCLUDE_IF
 
   /**
    * Creates a CoreController instance
@@ -424,9 +520,6 @@ class Engine {
     const qrKeyringBuilder = () => new QRHardwareKeyring();
     qrKeyringBuilder.type = QRHardwareKeyring.type;
 
-    const ledgerKeyringBuilder = () => new LedgerKeyring();
-    ledgerKeyringBuilder.type = LedgerKeyring.type;
-
     const keyringController = new KeyringController({
       removeIdentity: preferencesController.removeIdentity.bind(
         preferencesController,
@@ -455,8 +548,280 @@ class Engine {
         allowedActions: ['KeyringController:getState'],
       }),
       state: initialKeyringState || initialState.KeyringController,
-      keyringBuilders: [qrKeyringBuilder, ledgerKeyringBuilder],
+      // @ts-expect-error To Do: Update the type of QRHardwareKeyring to Keyring<Json>
+      keyringBuilders: [qrKeyringBuilder],
     });
+
+    ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+    /**
+     * Gets the mnemonic of the user's primary keyring.
+     */
+    const getPrimaryKeyringMnemonic = () => {
+      const [keyring]: any = keyringController.getKeyringsByType(
+        KeyringTypes.hd,
+      );
+      if (!keyring.mnemonic) {
+        throw new Error('Primary keyring mnemonic unavailable.');
+      }
+
+      return keyring.mnemonic;
+    };
+
+    const getAppState = () => {
+      const state = AppState.currentState;
+      return state === 'active';
+    };
+
+    const getSnapPermissionSpecifications = () => ({
+      ...buildSnapEndowmentSpecifications(Object.keys(ExcludedSnapEndowments)),
+      ...buildSnapRestrictedMethodSpecifications(
+        Object.keys(ExcludedSnapPermissions),
+        {
+          encrypt: encryptor.encrypt.bind(encryptor),
+          decrypt: encryptor.decrypt.bind(encryptor),
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+          // @ts-ignore
+          clearSnapState: this.controllerMessenger.call.bind(
+            this.controllerMessenger,
+            'SnapController:clearSnapState',
+          ),
+          getMnemonic: getPrimaryKeyringMnemonic.bind(this),
+          getUnlockPromise: getAppState.bind(this),
+          getSnap: this.controllerMessenger.call.bind(
+            this.controllerMessenger,
+            'SnapController:get',
+          ),
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+          // @ts-ignore
+          handleSnapRpcRequest: this.controllerMessenger.call.bind(
+            this.controllerMessenger,
+            'SnapController:handleRequest',
+          ),
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+          // @ts-ignore
+          getSnapState: this.controllerMessenger.call.bind(
+            this.controllerMessenger,
+            'SnapController:getSnapState',
+          ),
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+          // @ts-ignore
+          updateSnapState: this.controllerMessenger.call.bind(
+            this.controllerMessenger,
+            'SnapController:updateSnapState',
+          ),
+          maybeUpdatePhishingList: this.controllerMessenger.call.bind(
+            this.controllerMessenger,
+            'PhishingController:maybeUpdateState',
+          ),
+          isOnPhishingList: (origin: string) =>
+            this.controllerMessenger.call(
+              'PhishingController:testOrigin',
+              origin,
+            ).result,
+          showDialog: (
+            origin: string,
+            type: EnumToUnion<DialogType>,
+            content: any, // should be Component from '@metamask/snaps-ui';
+            placeholder?: any,
+          ) =>
+            approvalController.addAndShowApprovalRequest({
+              origin,
+              type,
+              requestData: { content, placeholder },
+            }),
+          showInAppNotification: (origin: string, args: NotificationArgs) => {
+            Logger.log(
+              'Snaps/ showInAppNotification called with args: ',
+              args,
+              ' and origin: ',
+              origin,
+            );
+          },
+        },
+      ),
+    });
+    ///: END:ONLY_INCLUDE_IF
+
+    const permissionController = new PermissionController({
+      messenger: this.controllerMessenger.getRestricted({
+        name: 'PermissionController',
+        allowedActions: [
+          `${approvalController.name}:addRequest`,
+          `${approvalController.name}:hasRequest`,
+          `${approvalController.name}:acceptRequest`,
+          `${approvalController.name}:rejectRequest`,
+          ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+          `SnapController:getPermitted`,
+          `SnapController:install`,
+          `SubjectMetadataController:getSubjectMetadata`,
+          ///: END:ONLY_INCLUDE_IF
+        ],
+      }),
+      state: initialState.PermissionController,
+      caveatSpecifications: getCaveatSpecifications({ getIdentities }),
+      // @ts-expect-error Inferred permission specification type is incorrect, fix after migrating to TypeScript
+      permissionSpecifications: {
+        ...getPermissionSpecifications({
+          getAllAccounts: () => keyringController.getAccounts(),
+        }),
+        ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+        ...getSnapPermissionSpecifications(),
+        ///: END:ONLY_INCLUDE_IF
+      },
+      unrestrictedMethods,
+    });
+
+    ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+    const subjectMetadataController = new SubjectMetadataController({
+      messenger: this.controllerMessenger.getRestricted({
+        name: 'SubjectMetadataController',
+        allowedActions: [`${permissionController.name}:hasPermissions`],
+      }),
+      state: initialState.SubjectMetadataController || {},
+      subjectCacheLimit: 100,
+    });
+
+    const setupSnapProvider = (snapId: string, connectionStream: Duplex) => {
+      Logger.log(
+        '[ENGINE LOG] Engine+setupSnapProvider: Setup stream for Snap',
+        snapId,
+      );
+      // TO DO:
+      // Develop a simpler getRpcMethodMiddleware object for SnapBridge
+      // Consider developing an abstract class to derived custom implementations for each use case
+      const bridge = new SnapBridge({
+        snapId,
+        connectionStream,
+        getRPCMethodMiddleware: ({ hostname, getProviderState }) =>
+          getRpcMethodMiddleware({
+            hostname,
+            getProviderState,
+            navigation: null,
+            getApprovedHosts: () => null,
+            setApprovedHosts: () => null,
+            approveHost: () => null,
+            title: { current: 'Snap' },
+            icon: { current: undefined },
+            isHomepage: () => false,
+            fromHomepage: { current: false },
+            toggleUrlModal: () => null,
+            wizardScrollAdjusted: { current: false },
+            tabId: false,
+            isWalletConnect: true,
+            isMMSDK: false,
+            url: { current: '' },
+            analytics: {},
+            injectHomePageScripts: () => null,
+          }),
+      });
+
+      bridge.setupProviderConnection();
+    };
+
+    const requireAllowlist = process.env.METAMASK_BUILD_TYPE === 'main';
+    const allowLocalSnaps = process.env.METAMASK_BUILD_TYPE === 'flask';
+
+    const snapsRegistryMessenger = this.controllerMessenger.getRestricted({
+      name: 'SnapsRegistry',
+      allowedEvents: [],
+      allowedActions: [],
+    });
+    const snapsRegistry = new JsonSnapsRegistry({
+      state: initialState.SnapsRegistry,
+      messenger: snapsRegistryMessenger,
+      refetchOnAllowlistMiss: requireAllowlist,
+      failOnUnavailableRegistry: requireAllowlist,
+      url: {
+        registry: 'https://acl.execution.metamask.io/latest/registry.json',
+        signature: 'https://acl.execution.metamask.io/latest/signature.json',
+      },
+      publicKey:
+        '0x025b65308f0f0fb8bc7f7ff87bfc296e0330eee5d3c1d1ee4a048b2fd6a86fa0a6',
+    });
+
+    this.snapExecutionService = new WebviewExecutionService({
+      messenger: this.controllerMessenger.getRestricted({
+        name: 'ExecutionService',
+      }),
+      setupSnapProvider: setupSnapProvider.bind(this),
+    });
+
+    const snapControllerMessenger = this.controllerMessenger.getRestricted({
+      name: 'SnapController',
+      allowedEvents: [
+        'ExecutionService:unhandledError',
+        'ExecutionService:outboundRequest',
+        'ExecutionService:outboundResponse',
+        'SnapController:snapInstalled',
+        'SnapController:snapUpdated',
+      ],
+      allowedActions: [
+        `${approvalController.name}:addRequest`,
+        `${permissionController.name}:getEndowments`,
+        `${permissionController.name}:getPermissions`,
+        `${permissionController.name}:hasPermission`,
+        `${permissionController.name}:hasPermissions`,
+        `${permissionController.name}:requestPermissions`,
+        `${permissionController.name}:revokeAllPermissions`,
+        `${permissionController.name}:revokePermissions`,
+        `${permissionController.name}:revokePermissionForAllSubjects`,
+        `${permissionController.name}:getSubjectNames`,
+        `${permissionController.name}:updateCaveat`,
+        `${approvalController.name}:addRequest`,
+        `${approvalController.name}:updateRequestState`,
+        `${permissionController.name}:grantPermissions`,
+        `${subjectMetadataController.name}:getSubjectMetadata`,
+        `${phishingController.name}:maybeUpdateState`,
+        `${phishingController.name}:testOrigin`,
+        `${snapsRegistry.name}:get`,
+        `${snapsRegistry.name}:getMetadata`,
+        `${snapsRegistry.name}:update`,
+        'ExecutionService:executeSnap',
+        'ExecutionService:terminateSnap',
+        'ExecutionService:terminateAllSnaps',
+        'ExecutionService:handleRpcRequest',
+      ],
+    });
+
+    const snapController = new SnapController({
+      environmentEndowmentPermissions: Object.values(EndowmentPermissions),
+      featureFlags: {
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-ignore
+        requireAllowlist,
+      },
+      state: initialState.SnapController || undefined,
+      messenger: snapControllerMessenger as any,
+      detectSnapLocation: (
+        location: string | URL,
+        options?: DetectSnapLocationOptions,
+      ) =>
+        detectSnapLocation(location, {
+          ...options,
+          allowLocal: allowLocalSnaps,
+          fetch: fetchFunction,
+        }),
+    });
+
+    this.controllerMessenger.subscribe(
+      'SnapController:snapAdded',
+      (snap: Snap, svgIcon: any = null) => {
+        const {
+          manifest: { proposedName },
+          version,
+        } = snap;
+        subjectMetadataController.addSubjectMetadata({
+          subjectType: SubjectType.Snap,
+          name: proposedName,
+          origin: snap.id,
+          version,
+          svgIcon,
+        });
+      },
+    );
+    ///: END:ONLY_INCLUDE_IF
+
+    const codefiTokenApiV2 = new CodefiTokenPricesServiceV2();
 
     const controllers = [
       keyringController,
@@ -522,6 +887,7 @@ class Engine {
           ),
         getOpenSeaApiKey: () => nftController.openSeaApiKey,
         addNft: nftController.addNft.bind(nftController),
+        getNftApi: nftController.getNftApi.bind(nftController),
         getNftState: () => nftController.state,
       }),
       currencyRateController,
@@ -551,7 +917,8 @@ class Engine {
         chainId: networkController.state.providerConfig.chainId,
         ticker: networkController.state.providerConfig.ticker ?? 'ETH',
         selectedAddress: preferencesController.state.selectedAddress,
-        coinGeckoHeader: process.env.COIN_GECKO_HEADER as string,
+        tokenPricesService: codefiTokenApiV2,
+        interval: 30 * 60 * 1000,
       }),
       new TransactionController({
         blockTracker:
@@ -614,29 +981,7 @@ class Engine {
       ),
       gasFeeController,
       approvalController,
-      new PermissionController({
-        messenger: this.controllerMessenger.getRestricted({
-          name: 'PermissionController',
-          allowedActions: [
-            `${approvalController.name}:addRequest`,
-            `${approvalController.name}:hasRequest`,
-            `${approvalController.name}:acceptRequest`,
-            `${approvalController.name}:rejectRequest`,
-          ],
-        }),
-        state: initialState.PermissionController,
-        caveatSpecifications: getCaveatSpecifications({ getIdentities }),
-        // @ts-expect-error Inferred permission specification type is incorrect, fix after migrating to TypeScript
-        permissionSpecifications: {
-          ...getPermissionSpecifications({
-            getAllAccounts: () => keyringController.getAccounts(),
-          }),
-          /*
-            ...this.getSnapPermissionSpecifications(),
-            */
-        },
-        unrestrictedMethods,
-      }),
+      permissionController,
       new SignatureController({
         messenger: this.controllerMessenger.getRestricted({
           name: 'SignatureController',
@@ -675,6 +1020,10 @@ class Engine {
         }),
         state: initialState.LoggingController,
       }),
+      ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+      snapController,
+      subjectMetadataController,
+      ///: END:ONLY_INCLUDE_IF
     ];
 
     if (isBlockaidFeatureEnabled()) {
@@ -685,12 +1034,8 @@ class Engine {
           cdnBaseUrl: process.env.BLOCKAID_FILE_CDN as string,
           messenger: this.controllerMessenger.getRestricted({
             name: 'PPOMController',
+            allowedEvents: ['NetworkController:stateChange'],
           }),
-          onNetworkChange: (listener) =>
-            this.controllerMessenger.subscribe(
-              AppConstants.NETWORK_STATE_CHANGE_EVENT,
-              listener,
-            ),
           onPreferencesChange: (listener) =>
             preferencesController.subscribe(listener),
           provider: networkController.getProviderAndBlockTracker().provider,
@@ -702,14 +1047,16 @@ class Engine {
           securityAlertsEnabled:
             initialState.PreferencesController?.securityAlertsEnabled ?? false,
           state: initialState.PPOMController,
-          ppomInitialisationCallback: (): any => {
-            store.dispatch({
-              type: 'SET_PPOM_INITIALIZATION_COMPLETED',
-              ppomInitializationCompleted: true,
-            });
-          },
         });
         controllers.push(ppomController as any);
+        this.controllerMessenger.subscribe(
+          AppConstants.PPOM_INITIALISATION_STATE_CHANGE_EVENT,
+          (ppomInitializationStatus: PPOMInitialisationStatusType) => {
+            store.dispatch(
+              UpdatePPOMInitializationStatus(ppomInitializationStatus),
+            );
+          },
+        );
       } catch (e) {
         Logger.log(`Error initializing PPOMController: ${e}`);
         return;
@@ -1093,6 +1440,10 @@ export default {
       TokensController,
       TokenDetectionController,
       NftDetectionController,
+      ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+      SnapController,
+      SubjectMetadataController,
+      ///: END:ONLY_INCLUDE_IF
       PermissionController,
       ApprovalController,
       LoggingController,
@@ -1128,6 +1479,10 @@ export default {
       GasFeeController,
       TokenDetectionController,
       NftDetectionController,
+      ///: BEGIN:ONLY_INCLUDE_IF(snaps)
+      SnapController,
+      SubjectMetadataController,
+      ///: END:ONLY_INCLUDE_IF
       PermissionController,
       ApprovalController,
       LoggingController,